feat: add createSerializer
