+++ index.js
@@ -1,19 +1,3 @@
-var toDiffableHtml = require('@algolia/diffable-html');
+var createSerializer = require('./createSerializer');
 
-module.exports = {
-  test(object) {
-    if (typeof object !== 'string') {
-      return false;
-    }
-
-    const trimmed = object.trim();
-    return (
-      trimmed.length > 2 &&
-      trimmed[0] === '<' &&
-      trimmed[trimmed.length - 1] === '>'
-    );
-  },
-  print(val) {
-    return toDiffableHtml(val).trim();
-  },
-};
+module.exports = createSerializer();
