+++ package.json
@@ -13,8 +13,7 @@
   },
   "files": [
     "index.js",
-    "asc.js",
-    "desc.js"
+    "createSerializer.js"
   ],
   "repository": {
     "type": "git",
@@ -38,7 +37,7 @@
   },
   "homepage": "https://github.com/rayrutjes/jest-serializer-html#readme",
   "dependencies": {
-    "@algolia/diffable-html": "^4.1.0"
+    "diffable-html": "^4.1.0"
   },
   "devDependencies": {
     "conventional-changelog-cli": "^1.3.4",
